+++ test/Unchase.Swashbuckle.AspNetCore.Extensions.Tests/Models/EnumWithMultipleValueLabels.cs
@@ -1,0 +1,23 @@
+﻿namespace Unchase.Swashbuckle.AspNetCore.Extensions.Tests.Models
+{
+    /// <summary>
+    /// Enum with multiple value labels.
+    /// </summary>
+    public enum EnumWithMultipleValueLabels
+    {
+        /// <summary>
+        /// Field 1.
+        /// </summary>
+        Field1 = 1,
+
+        /// <summary>
+        /// Another name for field 1.
+        /// </summary>
+        AnotherNameForField1 = Field1,
+
+        /// <summary>
+        /// Field 2.
+        /// </summary>
+        Field2 = 2,
+    }
+}